Add pricing functions (#33)
* Added module `Math/Stochastic` with functions `fapf` for transforming tree to asset pricing formula, `simplify` for reducing this (still WIP), `MathML.presentation` to render as XML.

* Fix the `American` combinator (again)

After discusison with @GeorgSchneider and reviewing the pricing
semantics, this should have the `Until` inside the `Anytime`.

The difference is evident when considering what happens when the `Until`
barrier is __after__ the lower bound of the `Anytime`. It will expire
before the first possible acquisition date.

* Refactor claims to consolidate 'asset' parameter

I've had to remove ObservableFunctions.daml for the time being.  This is
the module providing an implementation for Claim (->) t a The reason is
that now the first parameter is of kind * -> * -> * -> * (instead of the
previous * -> * -> * that satisfies (->)).

I'd like to maintain back-compat but not sure how; that needs more
thought and I'll leave it for another day.

* Rename module `FinancialClaim` to `Financial`

* Bump package version to 1.0.0

Co-authored-by: lJoublanc <luciano@joublanc.com>