remove docs
The docs are moving to the docs repo. I'm deleting on this side first,
before adding on the other side, because it seems better to not have the
docs available for modification for a little while than to have the docs
duplicated and having to make sure they stay in sync.